requirements.txt
anthropic>=0.39.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
python-docx>=0.8.11
flask>=3.0.0
